Refactor file structre - move things out of actionCreators/tags and into separate file & dir. Consolidate additional tags logic
